Support to block privilege escalation attempts and synchronise loading/unloading multiple eBPF programs
We block privilege escalation attempts by attaching a LSM hook to cred_prepare and checking if the
process is trying to escalate privileges using unshare.

Furthur in order to ensure proper reception of perf events and cleanup of code on exit, the loading
and unloading of eBPF programs has been synchronised using waitgroups.